Improve slow I/O behaviour
This adjusts the code to "lock" the UI when a slow I/O operation is
taking place on the background executor, and to display a progress
indicator to make it clear that something is happening. This adds an
interface for injecting delays into the I/O code for testing purposes.

Fix: https://github.com/io7m/jwheatsheaf/issues/3