Improve the filename field semantics
This attempts to improve the semantics of the filename field in file
choosers. The old behaviour was essentially to ignore the return key
when typing into the field. The new behaviour is to attempt to select
an item when pressing the return key and, if an item is selected and
the OK button is unlocked, focus is shifted to the OK button. The
user can then press return to confirm the selection.

Affects: https://github.com/io7m/jwheatsheaf/issues/28