Adjust click-to-select implementation
This commit contains a number of fixes. The implementation of file
selection modes [0] worked, but there was an interaction with code
that was written _after_ the PR was written but _before_ the PR
was merged. This change adjusts the behaviour to allow that code to
work properly.

A cancel() method was added to file choosers to allow for closing
them reliably. This is not expected to be useful outside of the test
suite, where ensuring that all file choosers are closed between each
test run is desirable.

This also makes a few adjustments to the test suite to reduce some
fragility.

[0]: https://github.com/io7m/jwheatsheaf/pull/27
Affects: https://github.com/io7m/jwheatsheaf/issues/28