Improve "select directly" behaviour
If the path entered into the "select directly" dialog is a
file, then set the current directory to the parent of the
file, and select the file.

If the path entered into the "select directly" dialog is a
directory, then set the current directory to the path,
and select ".".

Fix: https://github.com/io7m/jwheatsheaf/issues/29