Fix a silently-failing OK button
This fixes an issue where the "are all files filtered" check was
being carried out in the OK button method. This caused the button to
silently fail in CREATE mode when the name of a nonexistent file was
typed in and an "is regular file" check was provided for the file
selection mode.

Fix: https://github.com/io7m/jwheatsheaf/issues/34